Finalize creation of build operators function
